rasterize/pyproject.toml
[project]
name = "rasterize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cairosvg>=2.8.2",
    "pillow>=12.1.1",
]
